move:#15 AuthForm 컴포넌트 경로 변경 및 파일 구조 정리
