+++ src/app/(auth)/login/page.tsx
@@ -1,5 +1,5 @@
 import Link from 'next/link';
-import AuthForm from '@/components/features/auth-form';
+import AuthForm from '@/components/features/auth/AuthForm';
 import { PATH } from '@/constants/routerPath';
 
 const LoginPage = () => {
